ml-service/requirements.txt
# Core API
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0

# Object storage
minio==7.2.7

# --- Phase 2: Transcription ---
faster-whisper==1.0.1

# Audio processing
pydub==0.25.1

# --- Phase 3: Speech Analysis ---
librosa==0.10.2
praat-parselmouth==0.4.3

# --- Phase 3: NLP ---
spacy==3.7.4
nltk==3.8.1

# --- Phase 5: Embeddings ---
# sentence-transformers==2.7.0  # Heavy — uncomment when Phase 5 starts

# --- Optional: Pyannote speaker diarization ---
# Requires HUGGINGFACE_TOKEN env var and a Pyannote gated model agreement.
# pyannote.audio==3.1.1
